requirements.txt
cffi==1.10.0
setuptools==33.1.1
Flask==0.12
Flask_Babel==0.11.2
Flask-Login==0.4.0
Flask_Bcrypt==0.7.1
Flask-Themes2==0.1.4
Flask_WTF==0.14
WTForms==2.0.2
Werkzeug==0.10.4
sphinx_better_theme==0.13
linux_metrics==0.1.4
netifaces==0.10.4
Jinja2==2.8
six==1.10.0
chardet==2.3.0
lockfile==0.12.2
pip==9.0.1
pytz==2012c
requests==2.8.1
Twisted==15.5.0
Pygments==2.0.1
SQLAlchemy>=1.1
pyOpenSSL==17.0.0
python_dateutil==2.6.0
pyephem==3.7.6.0
flask_session==0.3.1
protobuf==3.3.0
python_magic==0.4.13
psutil==5.2.2
simplejson==3.10.0
sqlalchemy_utils==0.32.14
tornado==4.5.1
PyYAML==3.12
pyzmq==16.0.2
alembic==0.9.1
PyMySQL==0.7.11
gunicorn==19.7.1
cryptography==1.7
wit==4.2.0
imageio==2.2.0
